Move authorize to nspace repo temporarily
